Modified app and db files to use TS
