--- db.ts
@@ -1,14 +1,0 @@
-import mongoose from 'mongoose';
-
-const { MONGO_URI } = process.env;
-
-mongoose.connect(MONGO_URI!, {});
-
-const db = mongoose.connection;
-
-db.on('error', console.error.bind(console, 'MongoDB connection error: '));
-db.once('open', () => {
-    console.log('MongoDb connected');
-});
-
-export default db;
