+++ package.json
@@ -4,7 +4,7 @@
   "description": "<h1>Backend Analyst Candidate Test<",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon app.js",
+    "start:dev": "nodemon --exec ts-node src/app.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -21,6 +21,7 @@
     "@types/dotenv": "^8.2.0",
     "@types/express": "^4.17.21",
     "@types/node": "^20.11.19",
-    "nodemon": "^3.0.3"
+    "nodemon": "^3.0.3",
+    "ts-node": "^10.9.2"
   }
 }
